+++ migrations/0002_2018-04-27_root.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 INSERT INTO system (id, name) VALUES (1, 'rootSystem');
-SELECT setval('system_service.system_id_seq' :: regclass, 1);
+SELECT setval(concat(current_schema(), '.system_id_seq') :: regclass, 1);
 
 -- +goose Down
 DELETE FROM system