+++ repository/token.go
@@ -2,11 +2,13 @@ package repository
 
 import (
 	"context"
+	"database/sql"
 
 	"github.com/Masterminds/squirrel"
 	"github.com/integration-system/isp-kit/db"
 	"github.com/integration-system/isp-kit/db/query"
 	"github.com/pkg/errors"
+	"isp-system-service/domain"
 	"isp-system-service/entity"
 )
 
@@ -91,4 +93,28 @@ func (r Token) DeleteToken(ctx context.Context, tokens []string) (int, error) {
 	}
 
 	return int(rowsAffected), nil
+}
+
+func (r Token) AuthDataByToken(ctx context.Context, token string) (*entity.AuthData, error) {
+	q := `
+SELECT system_id, domain_id, service_id, app_id, token.expire_time, token.created_at
+FROM token
+         LEFT JOIN application
+                   ON token.app_id = application.id
+         LEFT JOIN service
+                   ON application.service_id = service.id
+         LEFT JOIN domain
+                   ON service.domain_id = domain.id
+WHERE token = $1
+`
+	result := entity.AuthData{}
+	err := r.db.SelectRow(ctx, &result, q, token)
+	if err != nil {
+		if err == sql.ErrNoRows {
+			return nil, domain.ErrTokenNotFound
+		}
+		return nil, errors.WithMessage(err, "select row db")
+	}
+
+	return &result, nil
 }